Move to `internal.annotation` from `server`
