+++ core/src/main/java/com/linecorp/armeria/internal/annotation/AnnotatedHttpServiceExtensions.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.server;
+package com.linecorp.armeria.internal.annotation;
 
 import static java.util.Objects.requireNonNull;
